% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditinal_prob.r
\name{cond_pr}
\alias{cond_pr}
\title{cond.pr}
\usage{
cond_pr(x, y, span = 0.5, bandwidth = NULL, rnn = NULL, knn = NULL)
}
\arguments{
\item{x}{vector or matrix whose rows contains predictors
of each observations.}

\item{y}{vector of the variable we want to estimate its condtional cdf}

\item{span}{fraction of total distance to be the bandwidth, between 0 and 1,
default to be 0.5 if bandwidth not supply.}

\item{bandwidth}{the bandwidth of x to evaluate the cdf,
overwrite the span input.}

\item{rnn}{fraction of total number of observations to be knn,
between 0 and 1. Set span to be NULL to use knn,
default to be 0.5, if not supply}

\item{knn}{the number of nearest neighbor to estimate the cdf,
should be less than the number of observation and greater than 0
(\code{knn < length(y)}.
setting this value overwrite the rnn}
}
\value{
vector of conditional probability \code{p(Y < y | x)}
}
\description{
Estimate conditional probability for each observation
Default to use local method over knn method,
if both arguement supply or unspecify, only the local method will use.
}
\examples{
cond_pr(1:5, 1:5) #[1] 0.33 0.5 0.6 0.75 1
cond_pr(1:5, 5:1, span=0) #[1] 1 1 1 1 1
cond_pr(1:5, 5:1, span=1) #[1] 1.0 0.8 0.6 0.4 0.2
cond_pr(1:5, 5:1, span=NULL) [1] 1. 0.66 0.66 0.66 0.33
cond_pr(1:5, 5:1, span=NULL, rnn=1) [1] 1.0 0.8 0.6 0.4 0.2
cond_pr(1:5, 5:1, span=NULL, rnn=0) [1] 1 1 1 1 1
cond_pr(1:5, 5:1, span=NULL, rnn=0, knn=1) [1] 1. 0.6 0.66 0.66 0.5
cond_pr(1:5, 5:1, span=NULL, rnn=0, knn=3) [1] 1.00 0.75 0.60 0.50 0.25
cond_pr(1:5, 5:1, bandwidth=1) [1] 1 0.6 0.6 0.6 0.5
cond_pr(matrix(1:12, nrow=6), 6:1) [1] 1. 0.75 0.6 0.6 0.5 0.33
cond_pr(matrix(1:12, nrow=6), 6:1, span = 0.6) [1] 1.0.8 0.67 0.5 0.4 0.25
cond_pr(matrix(1:12, nrow=6), 6:1, bandwidth = 5) 1. 0.8 0.6 0.5 0.4 0.25
cond_pr(matrix(1:12, nrow=6), 6:1, span = NULL) [1] 1. 0.75 0.6 0.6 0.5 0.25
cond_pr(matrix(1:12, nrow=6), 6:1, span = NULL, rnn = 0)
cond_pr(matrix(1:12, nrow=6), 6:1, span = NULL, rnn = 1)
cond_pr(matrix(1:12, nrow=6), 6:1, span = NULL, rnn = 1, knn=1)
}
