% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cq_reg.r
\name{cq.reg}
\alias{cq.reg}
\title{cq.reg}
\usage{
cq.reg(x, y, span = 0.5, bandwidth = NULL, rnn = NULL, knn = NULL)
}
\arguments{
\item{x}{vector or matrix whose rows contains predictors
of each observations.}

\item{y}{vector of the variable we want to estimate its condtional cdf}

\item{span}{fraction of total distance to be the bandwidth, between 0 and 1,
default to be 0.5 if bandwidth not supply.}

\item{bandwidth}{the bandwidth of x to evaluate the cdf,
overwrite the span input.}

\item{rnn}{fraction of total number of observations to be knn,
between 0 and 1. Set span to be NULL to use knn,
default to be 0.5, if not supply}

\item{knn}{the number of nearest neighbor to estimate the cdf,
should be less than the number of observation and greater than 0
(\code{knn < length(y)}.
setting this value overwrite the rnn}
}
\value{
a scam object
}
\description{
Estimate conditional quantile include two steps:
\enumerate{
\item estimate the conditional probability for each observation: tau=p(y|x)
Default to use local method over knn method,
if both arguement supply or unspecify,
only the local method will use.
\item apply scam package to fix x to y with one more predictor tau,
constrain tau to be monotonely nondecreasing.
}
}
